Add an IgnoreListMapper
There are occasions where errors have to be ignored such as
context cancelled by user request.

This adds a `IgnoreListMapper` which allow to specify a list of
maperr.Error which we want to ignore.
When this mapper is used in conjuntion with other mappers you can
ignore some error whilst map others.

Other changes:
- Added constructor for Hashable mapper along with Append so it has
an API more similar to the one used by the List mapper
- Migrated to go modules
- Added Hashable() to maperr.Error which return the Hashable error
for a given maperr.Error

Closes #13, #8

Satisfies JIRA Ticket: https://izettle.atlassian.net/browse/MSEU-497