Allow Map on list mapper to accept std library errors
Currently if Map is called on a ListMapper, and is not a maperr.Error
it will not map.

With this change it will cast the error to a maperr.Error and
map it using the maperr.Error.Equal() which will default to an
exact text value comparison for std errors.