Swaps the use of map.Error with std library errors
Since `maperr.Error` decorates `errors` is not a breaking change to
just use `errors`.

This change does that, and then does some casting within each function
that was expecting an `map.Error`.

This will simplify the use of the library and allow any mapper to work
with the `maperr.WithStatus` func

Other changes:
 - `errorWithStatus` and `formattedError` implement the `Unwrapper` interface
 - several types that had not been use anywhere except this library have
been unexported: `castError`, `newAppendStrategy`, `ignoreStrategy`,
`lastMapped`,
 - Added alias `LastMappedWithStatus` of `MappedWithStatus`