When mapping look for any error that match the first in the entire chain
The mappers are currenlty strict in the way they search for a mapped
error. They only look for the last one that was appended to the chain.

This can cause issues if withing the same layer new error are appended
and this breaks the mapping on the layer above.

This change ensures that when an error is mapped, the Map() func
sequentially looks for any error that match in the entire chain.

This is similar to what the stdlib `errors.As` and `errors.Is` do when unwrapping
errors.

Other changes:
 - Improved readme documenting the `map.WithStatus` usage