When an error is not found in the mappings default to server error
Currenlty when an error is not found in a mapping, it returns `nil`.
This means that the caller then is required to make an extra check
for `if err != nil` so can return a server error.

With this change, that extra check won't be required as if we have an
error that isn't mapped, is usually flagged a server error.

Satisfies JIRA Ticket: https://izettle.atlassian.net/browse/MSEU-1561